// yiqiheshebei/src/service/instrument-manage/damageRegistration/types.ts
export interface damageRegistrationList {
  total: number
  data: {
    id: number
    record_number: string
    cost: number
    total: number
    species: number
    create_time: string
    status: number
    desc: string
    damage_man: string
    repay_man: string
    repay_price: string
    repay_time: string
    user_name: string
    list: {
      id?: number
      name: string
      asset_number: string
      specification: string
      total: number
    }[]
  }[]
}
// ...
export interface instrumnetList {
  id: number
  asset_number: string
  bacth: string
  specification: string
  number: string
  name: string
  stockStatus: boolean
  unit_price: string
  picture: string
  weekList: []
  detail: {
    id: number
    unit_price: string
    instrumnet_id: number
    laboratory_name: string
    name: string
    specification: string
    number_plies: number
    store_id: number
    store_name: string
    total: number
  }[]
}
// ...
export interface addDamageRegistrationList {
  type: number
  status: number
  damage_id: number | string
  damage_man: string
  damage_desc: string
  repay_man: string
  repay_id: number | string
  repay_price: number | string
  repay_time: string
  repay_desc: string
  url: Array<string>
  list: {
    unit_price: string
    hardware_number: string
    asset_number: string
    name: string
    bacth: string
    specification: string
    laboratory_name: string
    store_name: string
    instrument_id: number
    laboratory_id: number
    store_id: number
    number_plies: number
    total: number
  }[]
}
// ...
export interface instrumentDetailList {
  status: number
  damage_man: string
  desc: string
  record_number: string
  create_time: string
  repay_man: string
  user_name: string
  repay_price: string
  repay_time: string
  repay_desc: string
  pic: {
    url_path: string
  }[]
  list: {
    id: number
    name: string
    asset_number: string
    laboratory_id: number
    store_id: number
    laboratory_name: string
    instrument_id: number
    store_name: string
    number_plies: number
    total: number
  }[]
}
// ...
export interface compensateInfoList {
  id: number
  type: number
  repay_man: string
  status: number
  damage_man: string
  repay_price: string
  repay_time: string
  repay_desc: string
  list: {
    id: number
    name: string
    asset_number: string
    laboratory_id: number
    store_id: number
    laboratory_name: string
    instrument_id: number
    store_name: string
    number_plies: number
    total: number
  }[]
}
// ...
export interface instrumentRegistrationList {
  total: number
  data: {
    id: number
    asset_number: string
    number: string
    name: string
    specification: string
    unit: string
    lend_num: number
    scrap_num: number
    maintenance_num: number
    damage_num: number
  }[]
}